Fix to pyflowgraph for input and output ports.
The ports that were being added through Kraken were always added as IO ports in the node when passed to the addPort method of the Node class.
KrakenInputPort and KrakenOutputPort now properly sub-class from the pyflowgraph InputPort and OutputPort objects and are added properly to each PortList.
Additionally new methods were added to find ports 'getInputPort()' and 'getOutputPort()' and are used when creating connections.
This now fixes the issue when both an input and an output have the same name.